backend/requirements/base.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-silk==5.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.6.0
gprof2dot==2025.4.14
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.8.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
PyPDF2==3.0.1
python-docx==1.0.1
nltk==3.8.1
